--- a/app/pocket.go
+++ b/app/pocket.go
@@ -46,11 +46,11 @@ type PocketCoreApp struct {
 }
 
 // new pocket core base
-func NewPocketBaseApp(logger log.Logger, db db.DB, cache bool, iavlCacheSize int64) *PocketCoreApp {
+func NewPocketBaseApp(logger log.Logger, db db.DB) *PocketCoreApp {
 	cdc = Codec()
 	bam.SetABCILogging(GlobalConfig.PocketConfig.ABCILogging)
 	// BaseApp handles interactions with Tendermint through the ABCI protocol
-	bApp := bam.NewBaseApp(appName, logger, db, cache, iavlCacheSize, auth.DefaultTxDecoder(cdc), cdc)
+	bApp := bam.NewBaseApp(appName, logger, db, auth.DefaultTxDecoder(cdc), cdc)
 	// set version of the baseapp
 	bApp.SetAppVersion(AppVersion)
 	// setup the key value store Keys