@jessicadaugherty 2022-10-04
@andrewnguyen22 I had closed this per the comment

> closes https://github.com/pokt-network/pocket-core/issues/1442

in favor of #1467

Lmk if that makes sense! CC @oten91